// Capstone Project/src/app/pages/landlord/host-post-management-page/screen/types.ts
export enum hostPostpageEnum {
  CREATE_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/CREATE_POST_HOST_REQUEST',
  CREATE_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/CREATE_POST_HOST_SUCCESS',
  CREATE_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/CREATE_POST_HOST_FAILURE',

  GET_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/GET_POST_HOST_REQUEST',
  GET_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/GET_POST_HOST_SUCCESS',
  GET_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/GET_POST_HOST_FAILURE',

  GET_POST_TYPE_HOST_REQUEST = 'APP/HOST_POST_PAGE/GET_POST_TYPE_HOST_REQUEST',
  GET_POST_TYPE_HOST_SUCCESS = 'APP/HOST_POST_PAGE/GET_POST_TYPE_HOST_SUCCESS',
  GET_POST_TYPE_HOST_FAILURE = 'APP/HOST_POST_PAGE/GET_POST_TYPE_HOST_FAILURE',

  EXTEND_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/EXTEND_POST_HOST_REQUEST',
  EXTEND_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/EXTEND_POST_HOST_SUCCESS',
  EXTEND_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/EXTEND_POST_HOST_FAILURE',

  VERIFY_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_REQUEST',
  VERIFY_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_SUCCESS',
  VERIFY_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_FAILURE',

  VERIFY_POST_HOST_AGAIN_REQUEST = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_AGAIN_REQUEST',
  VERIFY_POST_HOST_AGAIN_SUCCESS = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_AGAIN_SUCCESS',
  VERIFY_POST_HOST_AGAIN_FAILURE = 'APP/HOST_POST_PAGE/VERIFY_POST_HOST_AGAIN_FAILURE',

  SEARCH_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/SEARCH_POST_HOST_REQUEST',
  SEARCH_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/SEARCH_POST_HOST_SUCCESS',
  SEARCH_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/SEARCH_POST_HOST_FAILURE',

  DELETE_POST_HOST_REQUEST = 'APP/HOST_POST_PAGE/DELETE_POST_HOST_REQUEST',
  DELETE_POST_HOST_SUCCESS = 'APP/HOST_POST_PAGE/DELETE_POST_HOST_SUCCESS',
  DELETE_POST_HOST_FAILURE = 'APP/HOST_POST_PAGE/DELETE_POST_HOST_FAILURE',

  CLEAR_POST_HOST_STATE = 'APP/HOST/HOST_POST_PAGE/CLEAR_POST_HOST_STATE',

  CLEAR_MSG = 'APP/HOST/HOST_POST_PAGE/CLEAR_MSG',
}